pyproject.toml
[tool.poetry]
name = "cw5-skypro"
version = "0.1.0"
description = ""
authors = ["Mark Flamberg <riot.flamberg@gmail.com>"]
readme = "README.md"
packages = [{include = "cw5_skypro"}]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.24.1"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
